Add CI script for linting shell script
This will help catch any malformed shell scripts without having to manually run
them each time a change is made.

When/If we get to testing the Python script, we can simply change the language
in .travis.yml to Python and run the Python tests along with the install.sh and
build.sh scripts here.

https://github.com/caarlos0/shell-ci-build